refactor: extracted mui components overide to a new file
